NodeManager/ChildNode/requirements.txt
psutil
azure-common==1.1.28
azure-storage-common==2.1.0
azure-storage-file==2.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.2
dnspython
Flask==2.1.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
kafka-python==2.0.2
MarkupSafe==2.1.1
mongoengine==0.24.1
pycparser==2.21
pydantic==1.9.0
pymongo==4.0.2
python-dateutil==2.8.2
python-on-whales==0.41.0
python-dotenv
requests==2.27.1
six==1.16.0
tqdm==4.63.1
typer==0.4.1
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.7.0
